Extract compaction subsystem and remove legacy compact op
