+++ crates/alan/src/daemon/remote_control.rs
@@ -507,7 +507,7 @@ pub fn required_scope_for_op(op: &alan_protocol::Op) -> SessionScope {
     use alan_protocol::Op;
     match op {
         Op::Resume { .. } => SessionScope::Resume,
-        Op::Compact | Op::CompactWithOptions { .. } | Op::Rollback { .. } => SessionScope::Admin,
+        Op::CompactWithOptions { .. } | Op::Rollback { .. } => SessionScope::Admin,
         Op::Turn { .. }
         | Op::Input { .. }
         | Op::Interrupt
@@ -631,7 +631,10 @@ mod tests {
             }),
             SessionScope::Resume
         );
-        assert_eq!(required_scope_for_op(&Op::Compact), SessionScope::Admin);
+        assert_eq!(
+            required_scope_for_op(&Op::CompactWithOptions { focus: None }),
+            SessionScope::Admin
+        );
         assert_eq!(
             required_scope_for_op(&Op::CompactWithOptions {
                 focus: Some("preserve todos".to_string()),
