+++ crates/protocol/src/lib.rs
@@ -5,6 +5,7 @@
 //! various frontend interfaces (CLI, REST API, WebSocket).
 
 mod adaptive;
+mod compaction;
 mod content;
 mod event;
 mod op;
@@ -13,6 +14,11 @@ pub use adaptive::{
     AdaptiveForm, AdaptivePresentationHint, AdaptiveYieldCapabilities, ClientCapabilities,
     ConfirmationYieldPayload, CustomYieldPayload, DynamicToolYieldPayload, StructuredInputKind,
     StructuredInputOption, StructuredInputQuestion, StructuredInputYieldPayload,
+};
+pub use compaction::{
+    AppliedCompactionOutcome, CompactionMode, CompactionOutcome, CompactionReason,
+    CompactionRequestMetadata, CompactionResult, CompactionSkipReason, CompactionTrigger,
+    FailedCompactionOutcome, SkippedCompactionOutcome,
 };
 pub use content::{ContentPart, parts_to_text};
 pub use event::{Event, EventEnvelope, ToolDecisionAudit, YieldKind};
